Added Redirect URL option after authentication.
Added a text field to options that specifies where to redirect the user after login. Defaults to wp_login_url() so no configuration is necessary. Redirect URL will default to home_url(), however if $_REQUEST['redirect_to'] or the WPCC redirect_to option are set it switches to those in that order.

Updated user prompt for redirect URL